+++ cosmos_server/src/structure/planet/biosphere/mod.rs
@@ -32,6 +32,7 @@ use crate::{
 
 use super::generation::planet_generator::check_needs_generated_system;
 
+pub mod biosphere_generation;
 pub mod grass_biosphere;
 pub mod test_all_stone_biosphere;
 
@@ -46,6 +47,9 @@ struct NeedsBiosphereEvent {
 pub trait TGenerateChunkEvent {
     /// Creates the generate chunk event
     fn new(x: usize, y: usize, z: usize, structure_entity: Entity) -> Self;
+
+    fn get_structure_entity(&self) -> Entity;
+    fn get_chunk_coordinates(&self) -> (usize, usize, usize);
 }
 
 /// This has to be redone.
